+++ test/Conversion/ForthToMemRef/do-loop.mlir
@@ -7,9 +7,11 @@
 
 // Stack allocation and push 10, 0:
 // CHECK: %[[ALLOCA:.*]] = memref.alloca() : memref<256xi64>
-// CHECK: arith.constant 10 : i64
+// CHECK: arith.constant 10 : i32
+// CHECK: arith.extsi %{{.*}} : i32 to i64
 // CHECK: memref.store
-// CHECK: arith.constant 0 : i64
+// CHECK: arith.constant 0 : i32
+// CHECK: arith.extsi %{{.*}} : i32 to i64
 // CHECK: memref.store
 
 // Pop start and limit from stack:
@@ -42,8 +44,8 @@
 module {
   func.func private @main() {
     %0 = forth.stack !forth.stack
-    %1 = forth.constant %0(10 : i64) : !forth.stack -> !forth.stack
-    %2 = forth.constant %1(0 : i64) : !forth.stack -> !forth.stack
+    %1 = forth.constant %0(10 : i32) : !forth.stack -> !forth.stack
+    %2 = forth.constant %1(0 : i32) : !forth.stack -> !forth.stack
     %output_stack, %value = forth.pop %2 : !forth.stack -> !forth.stack, i64
     %output_stack_0, %value_1 = forth.pop %output_stack : !forth.stack -> !forth.stack, i64
     %alloca = memref.alloca() : memref<1xi64>
